api/requirements.txt
pymongo==3.12.3
django
djangorestframework
django-cors-headers
djongo
dnspy